Add loading spinner during stimulate (#14)
During development feedback is instant, but when communicating with production on slower networks, we need to show some feedback about loading the server response.

This PR adds two function to set the cursor to load and regular. We then call that before and after the `stimulate` on `toggleStatus` and `toggleDutyType`.

Discovered this while making my StimulusReflex blog post and figured that it would look nice here.